Apply design to existing components.
