+++ public/index.stache
@@ -14,8 +14,9 @@
     <can-import from="bitcentive/components/main-nav/" />
     <can-import from="bitcentive/components/page-home/" />
     <can-import from="bitcentive/components/page-auth/" />
-    <can-import from="bitcentive/components/page-dashboard/" />
     <can-import from="bitcentive/components/page-four-oh-four/" />
+    <can-import from="bitcentive/pages/contributors.component"/>
+    <can-import from="bitcentive/pages/dashboard.component"/>
 
     <main-nav {(session)}="session" {(page)}="page" {(subpage)}="subpage"></main-nav>
 
@@ -28,6 +29,9 @@
       {{/case}}
       {{#case 'dashboard'}}
         <page-dashboard></page-dashboard>
+      {{/case}}
+      {{#case 'contributors'}}
+        <page-contributors/>
       {{/case}}
       {{#case 'four-oh-four'}}
         <page-four-oh-four></page-four-oh-four>
