+++ public/pages/dashboard.component
@@ -1,0 +1,29 @@
+<can-component tag="page-dashboard">
+  <style type="less">
+    div.container {
+      background: white;
+      padding-bottom: 16px;
+      padding-top: 16px;
+    }
+  </style>
+  <template>
+    <can-import from="bitcentive/components/select-contribution-month/"/>
+    <can-import from="bitcentive/components/contribution-month/"/>
+
+    <div class="container">
+      <bit-select-contribution-month {(selected-contribution-month-id)}="contributionMonthId"/>
+      {{#if contributionMonthId}}
+        <bit-contribution-month {(contribution-month-id)}="contributionMonthId"/>
+      {{/if}}
+    </div>
+  </template>
+  <script type="view-model">
+    import DefineMap from 'can-define/map/';
+
+    export default DefineMap.extend({
+      contributionMonthId: {
+        type: "string",
+      },
+    });
+  </script>
+</can-component>
